backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.30.6",
    "sqlalchemy==2.0.25",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "alembic==1.13.1",
    "pydantic[email]>=2.11.9",
    "bcrypt==4.0.1",
    "resend==2.13.1"
]
